+++ R/create-network.R
@@ -93,6 +93,12 @@ nc_plot_network <- function(.tbl,
         .fn_node_rename <- function(x) x
     assert_is_function(.fn_node_rename)
 
+    # TODO: Fix this to be tidier, there should be a better way to do it.
+    if (!requireNamespace("ggplot2", quietly = TRUE))
+        rlang::abort("Can't find ggplot2, please install it.")
+    if (!requireNamespace("tidygraph", quietly = TRUE))
+        rlang::abort("Can't find tidygraph, please install it.")
+
     graph_data_prep <- .tbl %>%
         nc_adjacency_graph(.graph = .graph) %>%
         tidygraph::as_tbl_graph() %>%