+++ tests/testthat/helpers-test-prep.R
@@ -1,2 +1,6 @@
-library(survival)
 suppressPackageStartupMessages(library(dplyr, quietly = TRUE))
+
+metabolite_network <- simulated_data %>%
+    select(matches("metabolite")) %>%
+    nc_create_network()
+