+++ tests/testthat/test-model-estimates.R
@@ -1,15 +1,10 @@
-context("Exposure-side multiple model computation.")
+context("Exposure and outcome multiple model estimates.")
 
-metabolite_network <- simulated_data %>%
-    select(matches("metabolite")) %>%
-    nc_create_network()
-
-test_that("model estimation works and results are output", {
+test_that("exposure side estimation outputs correctly", {
     exposure_estimates <- simulated_data %>%
         nc_exposure_estimates(
             .graph = metabolite_network,
             .exposure = "exposure",
-            .adjustment_vars = "age",
             .model_function = lm
         )
 
@@ -17,16 +12,51 @@ test_that("model estimation works and results are output", {
     expect_identical(class(exposure_estimates)[1], "tbl_df")
 })
 
-test_that("adjustment variable can be added", {
-    exposure_estimates <- simulated_data %>%
-        nc_exposure_estimates(
+test_that("outcome side estimation outputs correctly", {
+    outcome_estimates <- simulated_data %>%
+        # To speed it up a bit.
+        sample_n(500) %>%
+        nc_outcome_estimates(
             .graph = metabolite_network,
-            .exposure = "exposure",
+            .outcome = "case_status",
+            .model_function = glm,
+            .model_arg_list = list(family = binomial(link = "logit")),
+            .exponentiate = TRUE
+        )
+
+    expect_type(outcome_estimates, "list")
+    expect_identical(class(outcome_estimates)[1], "tbl_df")
+})
+
+test_that("one or more adjustment variable can be added", {
+    one_adj <- simulated_data %>%
+        .compute_model_estimates(
+            .graph = metabolite_network,
+            .external_var = "exposure",
+            .external_side = "exposure",
+            .adjustment_vars = "age",
             .model_function = lm
         )
 
-    expect_type(exposure_estimates, "list")
-    expect_identical(class(exposure_estimates)[1], "tbl_df")
+    expect_type(one_adj, "list")
+    expect_identical(class(one_adj)[1], "tbl_df")
+
+    two_adj <- simulated_data %>%
+        mutate(Random = rnorm(nrow(.))) %>%
+        .compute_model_estimates(
+            .graph = metabolite_network,
+            .external_var = "exposure",
+            .external_side = "exposure",
+            .adjustment_vars = c("age", "Random"),
+            .model_function = lm
+        )
+
+    expect_type(two_adj, "list")
+    expect_identical(class(two_adj)[1], "tbl_df")
+
+    adjusted_variables <- unique(two_adj$adjusted_vars)
+    adjusted_variables <- strsplit(adjusted_variables, ", ")[[1]]
+    expect_identical(adjusted_variables, c("age", "Random"))
 })
 
 test_that("assertion checks pass", {
@@ -93,20 +123,15 @@ test_that("missingness in data still provides results", {
     expect_identical(class(exposure_estimates)[1], "tbl_df")
 })
 
-test_that("adding more than one adjustment variable includes it in the model", {
-    two_adj_vars <- simulated_data %>%
-        mutate(Random = rnorm(nrow(.))) %>%
-        nc_exposure_estimates(
+test_that("computes when using survival::Surv and coxph", {
+    skip_if_not_installed("survival")
+    outcome_estimates <- simulated_data %>%
+        nc_outcome_estimates(
             .graph = metabolite_network,
-            .exposure = "exposure",
-            .adjustment_vars = c("age", "Random"),
-            .model_function = lm
+            .outcome = "survival::Surv(survival_time, case_status)",
+            .model_function = survival::coxph
         )
 
-    expect_type(two_adj_vars, "list")
-    expect_identical(class(two_adj_vars)[1], "tbl_df")
-
-    adjusted_variables <- unique(two_adj_vars$adjusted_vars)
-    adjusted_variables <- strsplit(adjusted_variables, ", ")[[1]]
-    expect_identical(adjusted_variables, c("age", "Random"))
+    expect_type(outcome_estimates, "list")
+    expect_identical(class(outcome_estimates)[1], "tbl_df")
 })